--- src/aioway/_torch/README.md
@@ -1,3 +1,0 @@
-# Torch utils
-
-A set of torch related utilities.
