--- src/aioway/_torch/__init__.py
@@ -1,6 +1,0 @@
-# Copyright (c) AIoWay Authors - All Rights Reserved
-
-"A collection of `torch` and `tensordict` utils."
-
-from .fake import *
-from .hooks import *
