+++ src/aioway/codecs/images.py
@@ -11,7 +11,7 @@ from torchvision import io as vio
 from torchvision.transforms import v2 as tt
 
 from aioway._schemas import Attr, AttrLike
-from aioway._torch import is_fake_mode_on, torch_set_fake_mode_func
+from aioway.modes import is_fake_mode_on, torch_set_fake_mode_func
 
 from ._bases import TorchCompatible
