+++ src/aioway/modes/__init__.py
@@ -1,6 +1,7 @@
 # Copyright (c) AIoWay Authors - All Rights Reserved
 
 from .aten import *
+from .fake import *
 from .hists import *
 from .modes import *
 from .toggles import *
