+++ tests/modes/test_history.py
@@ -3,8 +3,7 @@
 import pytest
 import torch
 
-from aioway._torch import fake_mode, is_fake_mode_on
-from aioway.modes import fake_fn, track_fn
+from aioway.modes import fake_fn, fake_mode, is_fake_mode_on, track_fn
 
 
 @pytest.fixture
