+++ tests/schemas/test_devices.py
